requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
anthropic==0.5.0
anyio==3.7.1
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==7.1.2
colorama==0.4.6
distro==1.8.0
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.1.3
filelock==3.13.1
Flask==1.1.4
Flask-Cors==3.0.10
Flask-JWT-Extended==4.2.3
flask-mongoengine==1.0.0
flask-restx==0.4.0
Flask-SQLAlchemy==2.5.1
flask-wtf==1.2.1
fsspec==2023.10.0
google-api-core==2.13.0
google-auth==2.23.4
google-cloud-aiplatform==1.36.1
google-cloud-bigquery==3.13.0
google-cloud-core==2.3.3
google-cloud-resource-manager==1.10.4
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpc-google-iam-v1==0.12.6
grpcio==1.59.2
grpcio-status==1.59.2
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
huggingface-hub==0.17.3
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
iniconfig==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
MarkupSafe==2.0.1
mongoengine==0.27.0
openai==1.2.0
packaging==23.2
pkgutil-resolve-name==1.3.10
pluggy==1.3.0
proto-plus==1.22.3
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic-core==2.10.1
PyJWT==2.8.0
pymongo==4.6.0
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
replicate==0.17.0
requests==2.31.0
rpds-py==0.12.0
rsa==4.9
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
tokenizers==0.14.1
tomli==2.0.1
tqdm==4.66.1
typing-extensions==4.8.0
urllib3==2.0.7
Werkzeug==1.0.1
wtforms==3.1.1
zipp==3.17.0
